Add CurrentmonthRecords component
