+++ src/components/UI/Records/features/CurrentMonthRecords/index.ts
@@ -1,0 +1,1 @@
+export { CurrentMonthRecords } from './CurrentMonthRecords';
